Use Linkedin API v2 to fetch user info
Linkedin has depreciate its v1 API; it will stop working on the 1st March
2019 (https://engineering.linkedin.com/blog/2018/12/developer-program-updates).

The change to the v2 was done for the authorization & access_token
endpoints in 8951d956baf1 but the endpoint to fetch the info was still v1.

The v2 endpoint return a very different payload; there is a lot less info
with the default permission and there isn't the email address anymore.
Therefore an extra request is done to get the user's email address and a
custom UserResponse has been written to support the complex JSON schema.

Note: I get some inspiration from https://github.com/laravel/socialite/pull/310/files